removes analysis subpackage. raw experiment data analysis routines have been moved to experiment class modules
